Bot library update

+++ src/main/java/com/omfgdevelop/jiratelegrambot/repository/TaskQueueRepository.java
@@ -35,7 +35,7 @@ public interface TaskQueueRepository extends JpaRepository<Task, Long> {
 
     @Query(value = "delete from tasks_queue t where t.telegram_id =?1 and (t.status=4 or t.status=5 or t.status=6)", nativeQuery = true)
     @Modifying
-    void deleteAllUncompleted(Integer id);
+    void deleteAllUncompleted(Long id);
 
     List<Task> findAllByTelegramId(Long telegramId);
 }
